Isolated tests for static pages
