+++ src/applications/static-pages/facilities/tests/BasicFacilityListWidget.unit.spec.jsx
@@ -2,14 +2,17 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 import { mockApiRequest } from 'platform/testing/unit/helpers';
-import { mockFacilityLocatorApiResponse } from './mockFacilitiesData';
+import {
+  mockWidgetFacilitiesList,
+  mockFacilityLocatorApiResponse,
+} from './mockFacilitiesData';
 
-import OtherFacilityListWidget from '../../facilities/OtherFacilityListWidget';
+import BasicFacilityListWidget from '../BasicFacilityListWidget';
 
-describe('facilities <OtherFacilityListWidget>', () => {
+describe('facilities <FacilityListWidget>', () => {
   it('should render loading', () => {
     const tree = shallow(
-      <OtherFacilityListWidget facilities="vha_663,nca_120" />,
+      <BasicFacilityListWidget facilities={mockWidgetFacilitiesList} />,
       {
         disableLifecycleMethods: true,
       },
@@ -23,13 +26,14 @@ describe('facilities <OtherFacilityListWidget>', () => {
     mockApiRequest(mockFacilityLocatorApiResponse);
 
     const tree = shallow(
-      <OtherFacilityListWidget facilities="vha_663,nca_120" />,
+      <BasicFacilityListWidget facilities={mockWidgetFacilitiesList} />,
     );
     tree.instance().request.then(() => {
       tree.update();
       expect(tree.find('LoadingIndicator').exists()).to.be.false;
 
-      const facilityName = tree.find('h3');
+      const facilityNameComponent = tree.find('FacilityTitle').dive();
+      const facilityName = facilityNameComponent.find('h3');
       expect(facilityName.text()).to.contain(
         'Pittsburgh VA Medical Center-University Drive',
       );
@@ -40,11 +44,16 @@ describe('facilities <OtherFacilityListWidget>', () => {
         'University Drive CPittsburgh, PA 15240-1003',
       );
 
-      const mainPhone = tree.find('.main-phone');
+      const facilityPhoneComponent = tree.find('FacilityPhone').dive();
+      const mainPhone = facilityPhoneComponent.find('.main-phone');
       expect(mainPhone.text()).to.contain('Main phone: 866-482-7488');
 
-      const facilityType = tree.find('.facility-type');
-      expect(facilityType.text()).to.contain('VA Medical Center (VAMC)');
+      const mentalHealthClinic = facilityPhoneComponent.find(
+        '.mental-health-clinic-phone',
+      );
+      expect(mentalHealthClinic.text()).to.contain(
+        'Mental health clinic: 412-360-6600',
+      );
       tree.unmount();
       done();
     });
